Fixed a couple of issues, this module is not yet in use so no problem.
git-svn-id: svn://svn.code.sf.net/p/chibios/svn/trunk@17309 27425a3e-05d8-49a3-a47f-9c15f0e5edd8

+++ os/se/include/seloops.h
@@ -69,7 +69,7 @@ extern "C" {
                           rtcnt_t tmo, uint32_t *valp);
   bool seRegWaitAllClear8X(volatile uint8_t *p, uint8_t mask,
                            rtcnt_t tmo, uint8_t *valp);
-  bool seRegWaitAllClear16X(volatile uint16_t *p, uint32_t mask,
+  bool seRegWaitAllClear16X(volatile uint16_t *p, uint16_t mask,
                             rtcnt_t tmo, uint16_t *valp);
   bool seRegWaitAllClear32X(volatile uint32_t *p, uint32_t mask,
                             rtcnt_t tmo, uint32_t *valp);
